Transition to intermediate representation
This is the first commit in the transition to an intermediate representation
that sits between YAML or markdown files and the Anki interface.

The setup is now that the intermediate interface can write to the Anki
collection. Then there are separate modules that write or read the intermediate
interface from/to a YAML/markdown file.