Correct file path for buddy files
Before, if the user didn't give a global path to the

The problem is that when the input file is given as a relative path, the Python
file object does not by itself contain the absolute path. Then later, Anki
automatically seems to change the current directory, so then there is no way to
find the path for the buddy files.

The solution here is to add an attribute `absolutepath` to the file object.